--- MyPromise.js
@@ -1,23 +1,0 @@
-class MyPromise {
-  constructor() {
-    
-  }
-  static resolve() {
-    
-  }
-  static reject() {
-    
-  }
-  static all() {
-    
-  }
-  static race() {
-    
-  }
-  then() {
-    
-  }
-  catch() {
-    
-  }
-}
